Start using `long` for symbol table offsets instead of `int`

+++ Ghidra/Features/FileFormats/src/main/java/ghidra/file/formats/ios/ExtractedMacho.java
@@ -98,7 +98,7 @@ public class ExtractedMacho {
 					if (cmd instanceof SymbolTableCommand symbolTable) {
 						symbolTable.addSymbols(getExtraSymbols());
 					}
-					int offset = cmd.getLinkerDataOffset();
+					long offset = cmd.getLinkerDataOffset();
 					int size = cmd.getLinkerDataSize();
 					if (offset == 0 || size == 0) {
 						continue;
